Add 10 minute sleep to infinite loops
Several SoxBot scripts are using `while(true)` loops around their
business logic. This means that the scripts will do their work on the
wikis and then immediately start over to do the work again. The bot
tasks are important, but not so important that they should stress the
Tool Labs servers by spinning in tight infinite loops.

This patch introduces a 10 minute sleep to each `while(true)` loop to
reduce the frequency of operations from "as fast as the hardware and
software will support" to "approximately 144 times per day".